share SqlEnclaveAttestationParameters.Crypto and add NotSupported version for netstandard2.0
